fix duplicate valid_for_rib impls
