+++ mg-common/src/lib.rs
@@ -12,6 +12,8 @@ pub mod thread;
 
 use std::time::Duration;
 
+use oxnet::{IPV4_NET_WIDTH_MAX, IPV6_NET_WIDTH_MAX, IpNet, Ipv4Net, Ipv6Net};
+
 /// Format a Duration for human-readable display.
 ///
 /// Examples: "426d 13h 24m 9s", "3h 12m 5s", "1m 30s 250ms", "500ms"
@@ -166,4 +168,46 @@ macro_rules! write_lock {
     ($rwl:expr) => {
         $rwl.write().expect("rwlock write")
     };
+}
+
+pub trait IpNetExt {
+    fn valid_for_rib(&self) -> bool;
+}
+
+impl IpNetExt for Ipv4Net {
+    /// Check if a prefix contains a subnet that is valid for use in the RIB.
+    /// Currently this only checks if the prefix overlaps with Loopback
+    /// (127.0.0.0/8) or Multicast (224.0.0.0/4) address space. We deliberately
+    /// do not flag Class E (240.0.0.0/4) or Link-Local (169.254.0.0/16)
+    /// ranges as invalid, as some networks have deployed these as if they were
+    /// standard routable unicast addresses, which we need to handle.
+    fn valid_for_rib(&self) -> bool {
+        !(self.addr().is_loopback()
+            || self.addr().is_multicast()
+            || (self.addr().is_unspecified()
+                && self.width() == IPV4_NET_WIDTH_MAX))
+    }
+}
+
+impl IpNetExt for Ipv6Net {
+    /// Check if a prefix contains a subnet that is valid for use in the RIB.
+    /// Currently this only checks if the prefix carries the Unspecified or
+    /// Loopback address (::/128 or ::1/128), Multicast (ff00::/8) or Link-Local
+    /// Unicast (fe80::/10) address spaces.
+    fn valid_for_rib(&self) -> bool {
+        !(self.addr().is_loopback()
+            || self.addr().is_multicast()
+            || self.addr().is_unicast_link_local()
+            || (self.addr().is_unspecified()
+                && self.width() == IPV6_NET_WIDTH_MAX))
+    }
+}
+
+impl IpNetExt for IpNet {
+    fn valid_for_rib(&self) -> bool {
+        match self {
+            IpNet::V4(net4) => net4.valid_for_rib(),
+            IpNet::V6(net6) => net6.valid_for_rib(),
+        }
+    }
 }
